Implement user ad interaction page
