+++ frontend/src/components/AdPreferencesForm.jsx
@@ -1,32 +1,34 @@
-import { useState, useEffect } from 'react';
-import { Box, CheckboxGroup, Checkbox, Stack, Button, Heading } from '@chakra-ui/react';
+import React, { useEffect, useState } from 'react';
+import { Box, CheckboxGroup, Checkbox, Stack, Button, Heading, useToast } from '@chakra-ui/react';
 import '../styles/AdPreferencesForm.css';
-import { updateAdPreferences } from '../api/ads';
-import { defaultCategories } from '../utils/ads';
+import { updateAdPreferences } from '../api/ads.js';
+import { defaultCategories } from '../utils/ads.js';
 
 export default function AdPreferencesForm({ initial = [], onSave }) {
   const [values, setValues] = useState(initial);
+  const toast = useToast();
 
   useEffect(() => {
     setValues(initial);
   }, [initial]);
 
   const handleSave = async () => {
-    const res = await updateAdPreferences(values);
-    if (onSave) onSave(res.preferences);
+    try {
+      const res = await updateAdPreferences(values);
+      if (onSave) onSave(res.preferences);
+      toast({ title: 'Preferences saved', status: 'success' });
+    } catch (err) {
+      toast({ title: 'Failed to save preferences', status: 'error' });
+    }
   };
 
   return (
     <Box className="ad-preferences-form" p={4} borderWidth="1px" borderRadius="md">
-      <Heading size="sm" mb={2}>
-        Ad Preferences
-      </Heading>
+      <Heading size="sm" mb={2}>Ad Preferences</Heading>
       <CheckboxGroup value={values} onChange={setValues}>
         <Stack spacing={1}>
           {defaultCategories.map((cat) => (
-            <Checkbox key={cat} value={cat}>
-              {cat}
-            </Checkbox>
+            <Checkbox key={cat} value={cat}>{cat}</Checkbox>
           ))}
         </Stack>
       </CheckboxGroup>
