+++ frontend/src/styles/AdCard.css
