+++ frontend/src/styles/SharedUserInteractionPage.css
