+++ frontend/src/utils/ads.js
